refactor: extract formatting check
