+++ src/check/mod.rs
@@ -6,9 +6,10 @@ use solang_parser::pt::{
     ContractPart, FunctionAttribute, FunctionDefinition, FunctionTy, SourceUnitPart,
     VariableAttribute, VariableDefinition, Visibility,
 };
-use std::{error::Error, ffi::OsStr, fs, path::Path, process};
+use std::{error::Error, ffi::OsStr, fs, path::Path};
 use walkdir::WalkDir;
 
+pub mod check_formatting;
 pub mod report;
 
 // A regex matching valid test names, see the `validate_test_names_regex` test for examples.
@@ -24,7 +25,7 @@ static RE_VALID_CONSTANT_NAME: Lazy<Regex> =
 /// TODO
 pub fn run(taplo_opts: taplo::formatter::Options) -> Result<(), Box<dyn Error>> {
     let valid_names = validate_conventions();
-    let valid_fmt = validate_fmt(taplo_opts);
+    let valid_fmt = check_formatting::run(taplo_opts);
 
     if valid_names.is_ok() && valid_fmt.is_ok() {
         Ok(())
@@ -38,30 +39,6 @@ pub fn run(taplo_opts: taplo::formatter::Options) -> Result<(), Box<dyn Error>> 
 // =============================
 
 // -------- Top level validation methods --------
-
-fn validate_fmt(taplo_opts: taplo::formatter::Options) -> Result<(), Box<dyn Error>> {
-    // Check Solidity with `forge fmt`
-    let forge_status = process::Command::new("forge").arg("fmt").arg("--check").output()?;
-
-    // Print any warnings/errors from `forge fmt`.
-    let stderr = String::from_utf8(forge_status.stderr)?;
-    let forge_ok = forge_status.status.success() && stderr.is_empty();
-    print!("{stderr}"); // Prints nothing if stderr is empty.
-
-    // Check TOML with `taplo fmt`
-    let config_orig = fs::read_to_string("./foundry.toml")?;
-    let config_fmt = taplo::formatter::format(&config_orig, taplo_opts);
-    let taplo_ok = config_orig == config_fmt;
-
-    if !forge_ok || !taplo_ok {
-        eprintln!(
-            "{}: Formatting validation failed, run `scopelint fmt` to fix",
-            "error".bold().red()
-        );
-        return Err("Invalid fmt found".into())
-    }
-    Ok(())
-}
 
 fn validate_conventions() -> Result<(), Box<dyn Error>> {
     let paths = ["./src", "./script", "./test"];
